Separate dual channel and pause recording services
