+++ plugin-flex-ts-template-v2/src/feature-library/pause-recording/helpers/PauseRecordingService.ts
@@ -7,32 +7,7 @@ export interface RecordingResponse {
   recording: FetchedRecording;
 }
 
-class RecordingService extends ApiService {
-  startDualChannelRecording = async (callSid: string): Promise<FetchedRecording> => {
-    return new Promise((resolve, reject) => {
-      const encodedParams: EncodedParams = {
-        callSid: encodeURIComponent(callSid),
-        Token: encodeURIComponent(this.manager.user.token),
-      };
-
-      this.fetchJsonWithReject<RecordingResponse>(
-        `${this.serverlessProtocol}://${this.serverlessDomain}/features/dual-channel-recording/flex/create-recording`,
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-          body: this.buildBody(encodedParams),
-        },
-      )
-        .then((resp: RecordingResponse) => {
-          resolve(resp.recording);
-        })
-        .catch((error) => {
-          console.log('Error starting dual channel recording', error);
-          reject(error);
-        });
-    });
-  };
-
+class PauseRecordingService extends ApiService {
   pauseCallRecording = async (callSid: string, pauseBehavior: string): Promise<FetchedRecording> => {
     return new Promise((resolve, reject) => {
       const encodedParams: EncodedParams = {
@@ -138,4 +113,4 @@ class RecordingService extends ApiService {
   };
 }
 
-export default new RecordingService();
+export default new PauseRecordingService();
